Remove redundant ModelProviderType
