+++ llm-service/app/services/models/__init__.py
@@ -39,9 +39,9 @@ from .embedding import Embedding
 from .llm import LLM
 from .providers import get_provider_class
 from .reranking import Reranking
-from ._model_source import ModelSource
+from ...config import ModelSource
 
-__all__ = ["Embedding", "LLM", "Reranking", "ModelSource", "get_model_source"]
+__all__ = ["Embedding", "LLM", "Reranking", "get_model_source"]
 
 
 def get_model_source() -> ModelSource:
