+++ llm-service/app/services/models/providers/caii.py
@@ -43,7 +43,7 @@ from llama_index.core.postprocessor.types import BaseNodePostprocessor
 from packaging.version import Version
 
 from ._model_provider import _ModelProvider
-from .._model_source import ModelSource
+from app.config import ModelSource
 from ...caii.caii import (
     get_caii_llm_models,
     get_caii_embedding_models,
