+++ llm-service/app/services/models/providers/openai.py
@@ -44,10 +44,9 @@ from llama_index.embeddings.openai import OpenAIEmbedding
 from llama_index.llms.openai import OpenAI
 
 from ._model_provider import _ModelProvider
-from .._model_source import ModelSource
 from ...caii.types import ModelResponse
 from ...llama_utils import completion_to_prompt, messages_to_prompt
-from ....config import settings
+from ....config import settings, ModelSource
 
 
 class OpenAiModelProvider(_ModelProvider):
