+++ llm-service/app/services/query/querier.py
@@ -53,7 +53,7 @@ from .agents.tool_calling_querier import (
 from .flexible_retriever import FlexibleRetriever
 from .multi_retriever import MultiSourceRetriever
 from ..metadata_apis.session_metadata_api import Session
-from ..models._model_source import ModelSource
+from ...config import ModelSource
 from ..models import get_model_source
 
 if TYPE_CHECKING:
